moved spotifyAPI, spotifyPlayer scope to Home Comp
